requirements.txt
streamlit>=1.30.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0